Started to modify the config to work with Csycn v5. I think
