+++ Plugin.cs
@@ -14,16 +14,16 @@ namespace RandomMoons
     /// Main plugin class
     /// </summary>
     [BepInPlugin(modGUID, modName, modVersion)]
-    [BepInDependency("LethalAPI.Terminal")] // Thunderstore mod id : LethalAPI-LethalAPI_Terminal-1.0.1
-    [BepInDependency("com.sigurd.csync")] // Thunderstore mod id : Sigurd-CSync-3.1.1
-    [BepInDependency("ainavt.lc.lethalconfig", BepInDependency.DependencyFlags.SoftDependency)] // Thunderstore mod id : AinaVT-LethalConfig-1.3.4
-    [BepInDependency("com.willis.lc.lethalsettings", BepInDependency.DependencyFlags.SoftDependency)] // Thunderstore mod id : willis81808-LethalSettings-1.4.0
+    [BepInDependency("LethalAPI.Terminal")] // TODO: update version Thunderstore mod id : LethalAPI-LethalAPI_Terminal-1.0.1
+    [BepInDependency("com.sigurd.csync")] // TODO : update version Thunderstore mod id : Sigurd-CSync-3.1.1
+    [BepInDependency("ainavt.lc.lethalconfig", BepInDependency.DependencyFlags.SoftDependency)] // TODO: update version Thunderstore mod id : AinaVT-LethalConfig-1.3.4
+    [BepInDependency("com.willis.lc.lethalsettings", BepInDependency.DependencyFlags.SoftDependency)] // TODO: update version Thunderstore mod id : willis81808-LethalSettings-1.4.0
     public class RandomMoons : BaseUnityPlugin
     {
         // Basic mod infos
         internal const string modGUID = "InnohVateur.RandomMoons";
         internal const string modName = "RandomMoons";
-        internal const string modVersion = "1.2.3";
+        internal const string modVersion = "1.3.0";
 
         // Harmony instance
         private readonly Harmony harmony = new Harmony(modGUID);
@@ -35,7 +35,8 @@ namespace RandomMoons
         private TerminalModRegistry Commands;
 
         // Config Instance
-        public static SyncConfig CustomConfig;
+        // TODO: change the way the config works
+        //public static SyncConfig CustomConfig;
 
         // Log Source Instance
         internal static ManualLogSource mls;
@@ -45,7 +46,7 @@ namespace RandomMoons
         {
             
             // Instantiates the config
-            CustomConfig = new SyncConfig(base.Config);
+            //CustomConfig = new SyncConfig(base.Config);
 
             // Instantiates the Plugin
             if (Instance == null)
