+++ RandomMoons.csproj
@@ -31,6 +31,9 @@
     <Reference Include="BepInEx">
       <HintPath>..\..\..\..\..\SteamLibrary\steamapps\common\Lethal Company\BepInEx\core\BepInEx.dll</HintPath>
     </Reference>
+    <Reference Include="com.sigurd.csync">
+      <HintPath>..\..\..\..\DEPENDENCIES\com.sigurd.csync.dll</HintPath>
+    </Reference>
     <Reference Include="LethalAPI.Terminal">
       <HintPath>..\..\..\..\DEPENDENCIES\LethalAPI.Terminal.dll</HintPath>
     </Reference>
