+++ Utils/States.cs
@@ -14,7 +14,7 @@ namespace RandomMoons.Utils
         public static bool confirmedAutostart = false; // Confirms startUponArriving (actually needed)
         public static bool exploreASAP = false; // If the ship should travel to a random moon upon finishing leaving the current level
         public static List<string> visitedMoons = new List<string>(); // Lists the moons visited by exploring during the current quota
-        public static readonly string[] vanillaMoons = { "Level1Experimentation", "Level2Assurance", "Level3Vow", "Level4March", "Level5Rend", "Level6Dine", "Level7Offense", "Level8Titan" }; // All vanilla moons scene names
+        public static readonly string[] vanillaMoons = { "Level1Experimentation", "Level2Assurance", "Level3Vow", "Level4March", "Level5Rend", "Level6Dine", "Level7Offense", "Level8Titan" }; // TODO: Add new moons added with V50.  All vanilla moons scene names 
         public static readonly int companyBuildingLevelID = 3; // Gordion (Company Building) level ID
         public static string lastVisitedMoon; // Last visited moon using explore command / ship autoExplore
     }
